Внёс ваши корректировки. Единсвенное, не выполнил правило пяти в классе Matrix. Сам я считаю, что из-за использования библиотеки std::vector как основы для моего класса, там используется и деструктор оттуда и коснтруктора. Когда я попробовал переопределить оператор присваивания перемещением, половина функций полетело, а тут 1000+ строк. Очень не хочется всё переделывать. Вот... Что скажете?

+++ Shonhage_Shtrassen/Matrix.cpp
@@ -8,7 +8,7 @@
 #include <vector>
 
 
-Matrix::Matrix(int num_rows, int num_columns)
+Matrix::Matrix(size_t num_rows, size_t num_columns)
 {
     if (num_rows <= 0 || num_columns <= 0)
     {
@@ -36,6 +36,7 @@ Matrix::Matrix(std::vector<double> vec)
         matrix.push_back({vec[i]});
     }
 }
+
 
 Matrix::Matrix(std::vector<std::vector<double>> matrix)
 {
@@ -211,8 +212,8 @@ Matrix Matrix::inverse_matrix_from_unit()
                 matrix_1.matrix[i] = vector_minus(matrix_1.matrix[i], vector_mult(matrix_1.matrix[i0], matrix_1.matrix[i][i0]));
             }
         }
-    }
-    return inverse_matrix;
+    }                           
+    return inverse_matrix;      
 }
 
 double Matrix::determinant_manual()
